requirements.txt
Flask==2.2.2
kaggle==1.6.8
pycryptodome==3.20.0
Werkzeug==2.2.2  # Ensure compatibility with Flask